[TASK] Speed optimization for VirgilAsymmetricCipher operations
Point of the optimization is to store Random Number Generator object as internal state.
So it can be used among all operations that is needed for it.

Also VirgilAsymmetricCipher can be reused by reseting new keys, previously exception was thrown if was attempt to set key to object that already has the key.